extract MediaWithAuthentication logic into custom hook
